Fix issue with logging output arriving late
